+++ tests/testthat/test-autoplot-waveband.R
@@ -6,6 +6,22 @@ test_that("waveband", {
   set_annotations_default()
   vdiffr::expect_doppelganger("waveband-default-cie",
                               autoplot(CIE(), range = c(230, 430)))
+  vdiffr::expect_doppelganger("waveband-default-cie-wl-range-shrink",
+                              autoplot(CIE(),
+                                       w.length = 200:400,
+                                       range = c(230, 380)))
+  vdiffr::expect_doppelganger("waveband-cie-range-shrink-long",
+                              autoplot(CIE(),
+                                       w.length = 200:400,
+                                       range = c(230, 300, 380)))
+  # vdiffr::expect_doppelganger("waveband-default-cie-wl-range-expand",
+  #                             autoplot(CIE(),
+  #                                      w.length = 250:350,
+  #                                      range = c(200, 400)))
+  # vdiffr::expect_doppelganger("waveband-default-cie-range-r",
+  #                             autoplot(CIE(), range = c(NA, 340)))
+  # vdiffr::expect_doppelganger("waveband-default-cie",
+  #                             autoplot(CIE(), range = c(250, NA)))
   vdiffr::expect_doppelganger("waveband-default-red",
                               autoplot(Red()))
   vdiffr::expect_doppelganger("waveband-default-ylim",